Moved tests requiring balancer imports to separate packages
