+++ xds/internal/xdsclient/xdsresource/unmarshal_cds_test.go
@@ -18,7 +18,6 @@
 package xdsresource
 
 import (
-	"encoding/json"
 	"regexp"
 	"strings"
 	"testing"
@@ -27,35 +26,24 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	_ "google.golang.org/grpc/balancer/roundrobin"
-	"google.golang.org/grpc/internal/balancer/stub"
 	"google.golang.org/grpc/internal/envconfig"
 	"google.golang.org/grpc/internal/pretty"
-	internalserviceconfig "google.golang.org/grpc/internal/serviceconfig"
 	"google.golang.org/grpc/internal/testutils"
 	"google.golang.org/grpc/internal/xds/matcher"
-	"google.golang.org/grpc/serviceconfig"
-	"google.golang.org/grpc/xds/internal/balancer/ringhash"
-	"google.golang.org/grpc/xds/internal/balancer/wrrlocality"
-	_ "google.golang.org/grpc/xds/internal/balancer/wrrlocality"
 	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource/version"
 	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
-	v3 "github.com/cncf/xds/go/xds/type/v3"
 	v3clusterpb "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	v3corepb "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	v3endpointpb "github.com/envoyproxy/go-control-plane/envoy/config/endpoint/v3"
 	v3aggregateclusterpb "github.com/envoyproxy/go-control-plane/envoy/extensions/clusters/aggregate/v3"
 	v3leastrequestpb "github.com/envoyproxy/go-control-plane/envoy/extensions/load_balancing_policies/least_request/v3"
 	v3ringhashpb "github.com/envoyproxy/go-control-plane/envoy/extensions/load_balancing_policies/ring_hash/v3"
-	v3roundrobinpb "github.com/envoyproxy/go-control-plane/envoy/extensions/load_balancing_policies/round_robin/v3"
-	v3wrrlocalitypb "github.com/envoyproxy/go-control-plane/envoy/extensions/load_balancing_policies/wrr_locality/v3"
 	v3tlspb "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
 	v3discoverypb "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 	v3matcherpb "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
-	"github.com/golang/protobuf/proto"
 	anypb "github.com/golang/protobuf/ptypes/any"
-	structpb "github.com/golang/protobuf/ptypes/struct"
 )
 
 const (
@@ -287,549 +275,6 @@ func (s) TestValidateCluster_Failure(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			if update, err := validateClusterAndConstructClusterUpdate(test.cluster); err == nil {
 				t.Errorf("validateClusterAndConstructClusterUpdate(%+v) = %v, wanted error", test.cluster, update)
-			}
-		})
-	}
-}
-
-func wrrLocality(m proto.Message) *v3wrrlocalitypb.WrrLocality {
-	return &v3wrrlocalitypb.WrrLocality{
-		EndpointPickingPolicy: &v3clusterpb.LoadBalancingPolicy{
-			Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-				{
-					TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-						TypedConfig: testutils.MarshalAny(m),
-					},
-				},
-			},
-		},
-	}
-}
-
-func wrrLocalityAny(m proto.Message) *anypb.Any {
-	return testutils.MarshalAny(wrrLocality(m))
-}
-
-type customLBConfig struct {
-	serviceconfig.LoadBalancingConfig
-}
-
-func (s) TestValidateCluster_Success(t *testing.T) {
-	const customLBPolicyName = "myorg.MyCustomLeastRequestPolicy"
-	stub.Register(customLBPolicyName, stub.BalancerFuncs{
-		ParseConfig: func(json.RawMessage) (serviceconfig.LoadBalancingConfig, error) {
-			return customLBConfig{}, nil
-		},
-	})
-
-	oldCustomLBSupport := envconfig.XDSCustomLBPolicy
-	envconfig.XDSCustomLBPolicy = true
-	defer func() {
-		envconfig.XDSCustomLBPolicy = oldCustomLBSupport
-	}()
-	tests := []struct {
-		name             string
-		cluster          *v3clusterpb.Cluster
-		wantUpdate       ClusterUpdate
-		wantLBConfig     *internalserviceconfig.BalancerConfig
-		customLBDisabled bool
-	}{
-		{
-			name: "happy-case-logical-dns",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_LOGICAL_DNS},
-				LbPolicy:             v3clusterpb.Cluster_ROUND_ROBIN,
-				LoadAssignment: &v3endpointpb.ClusterLoadAssignment{
-					Endpoints: []*v3endpointpb.LocalityLbEndpoints{{
-						LbEndpoints: []*v3endpointpb.LbEndpoint{{
-							HostIdentifier: &v3endpointpb.LbEndpoint_Endpoint{
-								Endpoint: &v3endpointpb.Endpoint{
-									Address: &v3corepb.Address{
-										Address: &v3corepb.Address_SocketAddress{
-											SocketAddress: &v3corepb.SocketAddress{
-												Address: "dns_host",
-												PortSpecifier: &v3corepb.SocketAddress_PortValue{
-													PortValue: 8080,
-												},
-											},
-										},
-									},
-								},
-							},
-						}},
-					}},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName,
-				ClusterType: ClusterTypeLogicalDNS,
-				DNSHostName: "dns_host:8080",
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happy-case-aggregate-v3",
-			cluster: &v3clusterpb.Cluster{
-				Name: clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_ClusterType{
-					ClusterType: &v3clusterpb.Cluster_CustomClusterType{
-						Name: "envoy.clusters.aggregate",
-						TypedConfig: testutils.MarshalAny(&v3aggregateclusterpb.ClusterConfig{
-							Clusters: []string{"a", "b", "c"},
-						}),
-					},
-				},
-				LbPolicy: v3clusterpb.Cluster_ROUND_ROBIN,
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, LRSServerConfig: ClusterLRSOff, ClusterType: ClusterTypeAggregate,
-				PrioritizedClusterNames: []string{"a", "b", "c"},
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happy-case-no-service-name-no-lrs",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-				},
-				LbPolicy: v3clusterpb.Cluster_ROUND_ROBIN,
-			},
-			wantUpdate: emptyUpdate,
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happy-case-no-lrs",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_ROUND_ROBIN,
-			},
-			wantUpdate: ClusterUpdate{ClusterName: clusterName, EDSServiceName: serviceName, LRSServerConfig: ClusterLRSOff},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happiest-case",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_ROUND_ROBIN,
-				LrsServer: &v3corepb.ConfigSource{
-					ConfigSourceSpecifier: &v3corepb.ConfigSource_Self{
-						Self: &v3corepb.SelfConfigSource{},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{ClusterName: clusterName, EDSServiceName: serviceName, LRSServerConfig: ClusterLRSServerSelf},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-circuitbreakers",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_ROUND_ROBIN,
-				CircuitBreakers: &v3clusterpb.CircuitBreakers{
-					Thresholds: []*v3clusterpb.CircuitBreakers_Thresholds{
-						{
-							Priority:    v3corepb.RoutingPriority_DEFAULT,
-							MaxRequests: wrapperspb.UInt32(512),
-						},
-						{
-							Priority:    v3corepb.RoutingPriority_HIGH,
-							MaxRequests: nil,
-						},
-					},
-				},
-				LrsServer: &v3corepb.ConfigSource{
-					ConfigSourceSpecifier: &v3corepb.ConfigSource_Self{
-						Self: &v3corepb.SelfConfigSource{},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{ClusterName: clusterName, EDSServiceName: serviceName, LRSServerConfig: ClusterLRSServerSelf, MaxRequests: func() *uint32 { i := uint32(512); return &i }()},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-ring-hash-lb-policy-with-default-config",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_RING_HASH,
-				LrsServer: &v3corepb.ConfigSource{
-					ConfigSourceSpecifier: &v3corepb.ConfigSource_Self{
-						Self: &v3corepb.SelfConfigSource{},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName, LRSServerConfig: ClusterLRSServerSelf,
-				LBPolicy: &ClusterLBPolicyRingHash{MinimumRingSize: defaultRingHashMinSize, MaximumRingSize: defaultRingHashMaxSize},
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: "ring_hash_experimental",
-				Config: &ringhash.LBConfig{
-					MinRingSize: 1024,
-					MaxRingSize: 4096,
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-ring-hash-lb-policy-with-none-default-config",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_RING_HASH,
-				LbConfig: &v3clusterpb.Cluster_RingHashLbConfig_{
-					RingHashLbConfig: &v3clusterpb.Cluster_RingHashLbConfig{
-						MinimumRingSize: wrapperspb.UInt64(10),
-						MaximumRingSize: wrapperspb.UInt64(100),
-					},
-				},
-				LrsServer: &v3corepb.ConfigSource{
-					ConfigSourceSpecifier: &v3corepb.ConfigSource_Self{
-						Self: &v3corepb.SelfConfigSource{},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName, LRSServerConfig: ClusterLRSServerSelf,
-				LBPolicy: &ClusterLBPolicyRingHash{MinimumRingSize: 10, MaximumRingSize: 100},
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: "ring_hash_experimental",
-				Config: &ringhash.LBConfig{
-					MinRingSize: 10,
-					MaxRingSize: 100,
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-ring-hash-lb-policy-configured-through-LoadBalancingPolicy",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LoadBalancingPolicy: &v3clusterpb.LoadBalancingPolicy{
-					Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-						{
-							TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-								TypedConfig: testutils.MarshalAny(&v3ringhashpb.RingHash{
-									HashFunction:    v3ringhashpb.RingHash_XX_HASH,
-									MinimumRingSize: wrapperspb.UInt64(10),
-									MaximumRingSize: wrapperspb.UInt64(100),
-								}),
-							},
-						},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName,
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: "ring_hash_experimental",
-				Config: &ringhash.LBConfig{
-					MinRingSize: 10,
-					MaxRingSize: 100,
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-wrrlocality-rr-child-configured-through-LoadBalancingPolicy",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LoadBalancingPolicy: &v3clusterpb.LoadBalancingPolicy{
-					Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-						{
-							TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-								TypedConfig: wrrLocalityAny(&v3roundrobinpb.RoundRobin{}),
-							},
-						},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName,
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name: "round_robin",
-					},
-				},
-			},
-		},
-		{
-			name: "happiest-case-with-custom-lb-configured-through-LoadBalancingPolicy",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LoadBalancingPolicy: &v3clusterpb.LoadBalancingPolicy{
-					Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-						{
-							TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-								TypedConfig: wrrLocalityAny(&v3.TypedStruct{
-									TypeUrl: "type.googleapis.com/myorg.MyCustomLeastRequestPolicy",
-									Value:   &structpb.Struct{},
-								}),
-							},
-						},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName,
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: wrrlocality.Name,
-				Config: &wrrlocality.LBConfig{
-					ChildPolicy: &internalserviceconfig.BalancerConfig{
-						Name:   "myorg.MyCustomLeastRequestPolicy",
-						Config: customLBConfig{},
-					},
-				},
-			},
-		},
-		{
-			name: "custom-lb-env-var-not-set-ignore-load-balancing-policy-use-lb-policy-and-enum",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_RING_HASH,
-				LbConfig: &v3clusterpb.Cluster_RingHashLbConfig_{
-					RingHashLbConfig: &v3clusterpb.Cluster_RingHashLbConfig{
-						MinimumRingSize: wrapperspb.UInt64(20),
-						MaximumRingSize: wrapperspb.UInt64(200),
-					},
-				},
-				LoadBalancingPolicy: &v3clusterpb.LoadBalancingPolicy{
-					Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-						{
-							TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-								TypedConfig: testutils.MarshalAny(&v3ringhashpb.RingHash{
-									HashFunction:    v3ringhashpb.RingHash_XX_HASH,
-									MinimumRingSize: wrapperspb.UInt64(10),
-									MaximumRingSize: wrapperspb.UInt64(100),
-								}),
-							},
-						},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName,
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: "ring_hash_experimental",
-				Config: &ringhash.LBConfig{
-					MinRingSize: 20,
-					MaxRingSize: 200,
-				},
-			},
-			customLBDisabled: true,
-		},
-		{
-			name: "load-balancing-policy-takes-precedence-over-lb-policy-and-enum",
-			cluster: &v3clusterpb.Cluster{
-				Name:                 clusterName,
-				ClusterDiscoveryType: &v3clusterpb.Cluster_Type{Type: v3clusterpb.Cluster_EDS},
-				EdsClusterConfig: &v3clusterpb.Cluster_EdsClusterConfig{
-					EdsConfig: &v3corepb.ConfigSource{
-						ConfigSourceSpecifier: &v3corepb.ConfigSource_Ads{
-							Ads: &v3corepb.AggregatedConfigSource{},
-						},
-					},
-					ServiceName: serviceName,
-				},
-				LbPolicy: v3clusterpb.Cluster_RING_HASH,
-				LbConfig: &v3clusterpb.Cluster_RingHashLbConfig_{
-					RingHashLbConfig: &v3clusterpb.Cluster_RingHashLbConfig{
-						MinimumRingSize: wrapperspb.UInt64(20),
-						MaximumRingSize: wrapperspb.UInt64(200),
-					},
-				},
-				LoadBalancingPolicy: &v3clusterpb.LoadBalancingPolicy{
-					Policies: []*v3clusterpb.LoadBalancingPolicy_Policy{
-						{
-							TypedExtensionConfig: &v3corepb.TypedExtensionConfig{
-								TypedConfig: testutils.MarshalAny(&v3ringhashpb.RingHash{
-									HashFunction:    v3ringhashpb.RingHash_XX_HASH,
-									MinimumRingSize: wrapperspb.UInt64(10),
-									MaximumRingSize: wrapperspb.UInt64(100),
-								}),
-							},
-						},
-					},
-				},
-			},
-			wantUpdate: ClusterUpdate{
-				ClusterName: clusterName, EDSServiceName: serviceName,
-			},
-			wantLBConfig: &internalserviceconfig.BalancerConfig{
-				Name: "ring_hash_experimental",
-				Config: &ringhash.LBConfig{
-					MinRingSize: 10,
-					MaxRingSize: 100,
-				},
-			},
-		},
-	}
-
-	oldAggregateAndDNSSupportEnv := envconfig.XDSAggregateAndDNS
-	envconfig.XDSAggregateAndDNS = true
-	defer func() { envconfig.XDSAggregateAndDNS = oldAggregateAndDNSSupportEnv }()
-	oldRingHashSupport := envconfig.XDSRingHash
-	envconfig.XDSRingHash = true
-	defer func() { envconfig.XDSRingHash = oldRingHashSupport }()
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if test.customLBDisabled {
-				envconfig.XDSCustomLBPolicy = false
-				defer func() {
-					envconfig.XDSCustomLBPolicy = true
-				}()
-			}
-			update, err := validateClusterAndConstructClusterUpdate(test.cluster)
-			if err != nil {
-				t.Errorf("validateClusterAndConstructClusterUpdate(%+v) failed: %v", test.cluster, err)
-			}
-			// Ignore the raw JSON string into the cluster update. JSON bytes
-			// are nondeterministic (whitespace etc.) so we cannot reliably
-			// compare JSON bytes in a test. Thus, marshal into a Balancer
-			// Config struct and compare on that. Only need to test this JSON
-			// emission here, as this covers the possible output space.
-			if diff := cmp.Diff(update, test.wantUpdate, cmpopts.EquateEmpty(), cmpopts.IgnoreFields(ClusterUpdate{}, "LBPolicy", "LBPolicyJSON")); diff != "" {
-				t.Errorf("validateClusterAndConstructClusterUpdate(%+v) got diff: %v (-got, +want)", test.cluster, diff)
-			}
-			bc := &internalserviceconfig.BalancerConfig{}
-			if err := json.Unmarshal(update.LBPolicyJSON, bc); err != nil {
-				t.Fatalf("failed to unmarshal JSON: %v", err)
-			}
-			if diff := cmp.Diff(bc, test.wantLBConfig); diff != "" {
-				t.Fatalf("update.LBConfig got unexpected output, diff (-got +want): %v", diff)
 			}
 		})
 	}
